move `src/ui/utils` to `src/utils/task`
